Update APIs Logic

+++ src/movie/movie/movie.service.ts
@@ -1,5 +1,6 @@
-import { MovieDto, CreateMovieDto, IMovieDto, UpdateMovieDto } from './models/movie.model';
-import { defaultIfEmpty, filter, from, map, Observable, tap } from 'rxjs';
+import { MovieCategory, MovieCategoryDocument } from '../moviecategory/moviecategory.schema';
+import { CreateMovieDto, UpdateMovieDto } from './models/movie.model';
+import { from, map, Observable, tap } from 'rxjs';
 import { Movie, MovieDocument } from './models/movie.schema';
 import { InjectModel } from '@nestjs/mongoose';
 import { AppService } from 'src/app.service';
@@ -8,21 +9,56 @@ import { Model } from 'mongoose';
 
 @Injectable()
 export class MovieService {
+  get movieQuery(): any[] {
+    return [
+      { $project: { isDeleted: 0, __v: 0 } },
+      {
+        $lookup: {
+          from: 'moviecategories',
+          localField: '_id',
+          foreignField: 'movie',
+          pipeline: [
+            { $project: { _id: 0, __v: 0, movie: 0 } },
+            {
+              $lookup: {
+                from: 'categories',
+                localField: 'category',
+                foreignField: '_id',
+                as: 'category',
+              },
+            },
+            { $unwind: { path: '$category' } },
+            {
+              $group: {
+                _id: '$category._id',
+                name: { $first: '$category.name' },
+                translationKey: { $first: '$category.translationKey' },
+                createdTime: { $first: '$category.createdTime' },
+                updatedTime: { $first: '$category.updatedTime' },
+              },
+            },
+          ],
+          as: 'categories',
+        },
+      },
+    ];
+  }
+
   constructor(
     @InjectModel(Movie.name) private readonly model: Model<MovieDocument>,
-    private readonly appService: AppService
+    @InjectModel(MovieCategory.name)
+    private readonly movieCategory: Model<MovieCategoryDocument>,
+    private readonly appService: AppService,
   ) {}
 
   /**
    *
    * @returns
    */
-  findAll(): Observable<IMovieDto[]> {
-    const query = this.model.find();
+  findAll(): Observable<Movie[]> {
+    const query = this.model.aggregate([{ $match: { isDeleted: false } }, ...this.movieQuery]);
 
-    return from(query.exec()).pipe(
-      map((movieList) => movieList.map((movie) => new MovieDto(movie).object)),
-    );
+    return from(query.exec());
   }
 
   /**
@@ -30,13 +66,19 @@ export class MovieService {
    * @param _id
    * @returns
    */
-  findById(_id: string): Observable<IMovieDto> {
-    const query = this.model.findOne({ _id });
+  findById(_id: string): Observable<Movie> {
+    const query = this.model.aggregate([
+      {
+        $match: {
+          $expr: { $eq: ['$_id', { $toObjectId: _id }] },
+          isDeleted: false,
+        },
+      },
+      { $project: { _id: 0 } },
+      ...this.movieQuery,
+    ]);
 
-    return from(query.exec()).pipe(
-      defaultIfEmpty(MovieDto.emptyObject),
-      map((movie) => new MovieDto(movie).object),
-    );
+    return from(query.exec()).pipe(map((movies) => movies[0] ?? null));
   }
 
   /**
@@ -44,10 +86,20 @@ export class MovieService {
    * @param createMovieDto
    * @returns
    */
-  create(createMovieDto: CreateMovieDto): Observable<IMovieDto> {
+  create(createMovieDto: CreateMovieDto): Observable<Movie> {
+    const { categories } = createMovieDto;
     const createdMovie = new this.model(createMovieDto);
 
-    return from(createdMovie.save()).pipe(map((movie) => new MovieDto(movie).object));
+    if (categories) {
+      this.movieCategory.insertMany(
+        categories.map((category) => ({
+          movie: createdMovie._id,
+          category: category,
+        })),
+      );
+    }
+
+    return from(createdMovie.save());
   }
 
   /**
@@ -56,21 +108,36 @@ export class MovieService {
    * @param updateMovieDto
    * @returns
    */
-  update(_id: string, updateMovieDto: UpdateMovieDto): Observable<IMovieDto> {
-    const { poster } = updateMovieDto;
+  update(_id: string, updateMovieDto: UpdateMovieDto): Observable<Movie> {
+    const { poster, categories, ...updatedData } = updateMovieDto;
+
+    if (poster) {
+      (updatedData as any).poster = poster;
+    }
+
     const query = this.model.findOneAndUpdate(
       { _id },
-      { $set: { ...updateMovieDto, updatedTime: Date.now() } },
+      { $set: { ...updatedData, updatedTime: Date.now() } },
     );
+    query.projection({ __v: 0, isDeleted: 0 });
+
+    if (categories) {
+      this.movieCategory.deleteMany({ movies: _id }, () => {
+        this.movieCategory.insertMany(
+          categories.map((category) => ({
+            category: category,
+            movie: _id,
+          })),
+        );
+      });
+    }
 
     return from(query.exec()).pipe(
-      defaultIfEmpty(MovieDto.emptyObject),
-      tap(movie => {
-        if(movie.poster !== poster && movie.poster) {
+      tap((movie) => {
+        if (poster && movie.poster !== poster && movie.poster) {
           this.appService.removeUploadImage(movie.poster as string);
         }
       }),
-      map((movie) => new MovieDto(movie).object),
     );
   }
 
@@ -80,8 +147,11 @@ export class MovieService {
    * @returns
    */
   delete(_id: string): Observable<boolean> {
-    const query = this.model.findOneAndUpdate({ _id }, { $set: { isDeleted: true } });
+    const query = this.model.findOneAndUpdate(
+      { _id },
+      { $set: { isDeleted: true } },
+    );
 
-    return from(query.exec()).pipe(map((movie) => movie.isDeleted));
+    return from(query.exec()).pipe(map((movie) => !movie.isDeleted));
   }
 }
